+++ crates/nostr/examples/nip19.rs
@@ -7,8 +7,13 @@ use nostr::prelude::*;
 fn main() -> Result<()> {
     let pubkey =
         PublicKey::from_hex("3bf0c63fcb93463407af97a5e5ee64fa883d107ef9e558472c4eb9aaaefa459d")?;
-    let profile =
-        Nip19Profile::new(pubkey, vec!["wss://r.x.com", "wss://djbas.sadkb.com"]).unwrap();
+    let profile = Nip19Profile::new(
+        pubkey,
+        [
+            RelayUrl::parse("wss://r.x.com").unwrap(),
+            RelayUrl::parse("wss://djbas.sadkb.com").unwrap(),
+        ],
+    );
     println!("{}", profile.to_bech32()?);
 
     Ok(())